flytt translate til locale.ts
