+++ server/src/site/ssr/htmlRenderer.ts
@@ -2,7 +2,7 @@ import { buildHtmlTemplate, getTemplateWithDecorator } from './templateBuilder';
 import { ViteDevServer } from 'vite';
 // @ts-ignore (this file is generated by vite build)
 import { render } from '../../_ssr-dist/main-server';
-import { translate } from '../../../../common/breadcrumbs';
+import { translate } from '../../../../common/locale';
 import { Locale, defaultLocale } from '../../../../common/locale';
 import { URLs } from '../../urls';
 
@@ -49,11 +49,7 @@ export const prodRender: HtmlRenderer = async (url) => {
     const template = await getTemplateWithDecorator(url);
 
     try {
-        //TODO bruk await? brukt i nav-office-search
         const appHtml = render(url);
-        //TODO fjern log
-        console.log(`App html: ${appHtml.slice(0, 200)}`);
-
         return processTemplate(template, appHtml, url);
     } catch (e) {
         console.error(`Rendering failed ${e}}`);
